requirements.txt
firebase-admin==6.2.0
google-generativeai==0.3.0
google-cloud-firestore==2.11.0
python-dotenv==1.0.0
requests==2.31.0
